netlify/functions/requirements.txt
langchain>=0.2.0
langchain-google-genai>=1.0.0
langchain-community
faiss-cpu
fastapi
uvicorn
tiktoken
google-generativeai 
